Refactorisation of api code
