+++ api/src/main/java/com/daar/mylibrary/MylibraryApplication.java
@@ -5,11 +5,12 @@ import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
 import io.swagger.v3.oas.annotations.info.Info;
 import io.swagger.v3.oas.annotations.security.SecurityScheme;
+import io.swagger.v3.oas.annotations.servers.Server;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@OpenAPIDefinition(info = @Info(title = "My Library API", version = "0.1", description = "Books information"))
+@OpenAPIDefinition(info = @Info(title = "My Library API", version = "0.1", description = "Books information"), servers = @Server(url = "/v1"))
 @SecurityScheme(name = "globalSecurity", scheme = "bearer", bearerFormat = "JWT", type = SecuritySchemeType.HTTP, in = SecuritySchemeIn.HEADER)
 public class MylibraryApplication {
