+++ api/src/main/java/com/daar/mylibrary/controller/AuthorsController.java
@@ -1,8 +1,8 @@
 package com.daar.mylibrary.controller;
 
-import com.daar.mylibrary.response.Authors.AuthorsResponse;
-import com.daar.mylibrary.response.ErrorResponse;
-import com.daar.mylibrary.response.Response;
+import com.daar.mylibrary.dto.response.Authors.AuthorsResponse;
+import com.daar.mylibrary.dto.response.ErrorResponse;
+import com.daar.mylibrary.dto.response.Response;
 import com.daar.mylibrary.service.AuthorsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
@@ -32,7 +32,7 @@ public class AuthorsController {
     @Autowired
     private AuthorsService authorsService;
 
-    @Operation(summary = "[User] Basic search of authors of the Gutenberg library", description = "Search books from `keyword` by passing the param `search`. \nYou can also filter the search by `type`, select an `algorithm` and paginate the results.\n ### Permissions needed to access resources : \n- read:books\n- update:books")
+    @Operation(summary = "[User] Basic search of authors of the Gutenberg library", description = "Search books from `keyword` by passing the param `search`. \nYou can also filter the search by `type`, select an `algorithm` and paginate the results.\n ### Permissions needed to access resources : \n- read:books\n- read:authors")
     @ApiResponse(responseCode = "200", description = "Author founded", content = { @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = AuthorsResponse.class))) })
     @ApiResponse(responseCode = "401", description = "The authentication or authorization failed", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "404", description = "Author not founded", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
