+++ api/src/main/resources/application.properties
@@ -1,13 +1,15 @@
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url=${MYSQL_URL:jdbc:mysql://localhost:3306/mylibrary}
+spring.datasource.url=${MYSQL_URI:jdbc:mysql://localhost:3306/mylibrary}
 spring.datasource.username=${MYSQL_USER:em}
 spring.datasource.password=${MYSQL_PASSWORD:daar2021}
 spring.mvc.static-path-pattern=/static/**
 spring.servlet.multipart.max-file-size=1000KB
 spring.servlet.multipart.max-request-size=1000KB
 spring.profiles.active=${ENV_PROFILE:dev}
-spring.elasticsearch.uris=${ELASTIC_ENDPOINT:localhost:9200}
+spring.elasticsearch.uris=${ELASTIC_URI:localhost:9200}
+spring.elasticsearch.username=${ELASTIC_USER:elastic}
+spring.elasticsearch.password=${ELASTIC_PASSWORD:root}
 springdoc.swagger-ui.disable-swagger-default-url=true
 springdoc.swagger-ui.use-root-path=true
 spring.security.oauth2.resourceserver.jwt.issuer-uri=https://dev-mhw-6lg5.us.auth0.com/
